put config load and build options into a single file
